Fix ambiguity by renaming AppDomain.Instance to AppDomainInstance
