+++ tracer/test/test-applications/regression/dependency-libs/AppDomainInstance/AppDomainInstance.csproj
