+++ tracer/test/test-applications/regression/dependency-libs/AppDomainInstance/AppDomainInstanceProgram.cs
@@ -7,7 +7,7 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Threading;
 
-namespace AppDomain.Instance
+namespace AppDomainInstance
 {
     public class AppDomainInstanceProgram : MarshalByRefObject
     {
