+++ tracer/test/test-applications/regression/dependency-libs/AppDomainInstance/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "AppDomain.Instance": {
+    "AppDomainInstance": {
       "commandName": "Project",
       "environmentVariables": {
         "COR_ENABLE_PROFILING": "0",
